requirements.txt
# requirements.txt
sqlalchemy~=2.0.25
psycopg2==2.9.10
python-dotenv==1.0.1
faker==25.2.0
pandas==2.2.3
codecarbon==2.8.3
matplotlib~=3.10.1
uuid~=1.30